+++ src/desktop/src/ui/Index.js
@@ -1,8 +1,4 @@
 /* global Electron */
-import head from 'lodash/head';
-import last from 'lodash/last';
-import split from 'lodash/split';
-import isString from 'lodash/isString';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
@@ -13,12 +9,11 @@ import { withTranslation } from 'react-i18next';
 
 import { parseAddress } from 'libs/iota/utils';
 import { ALIAS_MAIN } from 'libs/constants';
-import { fetchVersions } from 'libs/utils';
+import { fetchVersions, fetchIsSeedMigrationUp, VALID_IOTA_SUBDOMAIN_REGEX } from 'libs/utils';
 
 import { getAccountNamesFromState, isSettingUpNewAccount } from 'selectors/accounts';
 
 import { setOnboardingComplete, setAccountInfoDuringSetup } from 'actions/accounts';
-import { setViewingMoonpayPurchases } from 'actions/ui';
 import {
     setPassword,
     clearWalletData,
@@ -28,20 +23,12 @@ import {
     shouldUpdate,
     forceUpdate,
     displayTestWarning,
+    displaySeedMigrationAlert,
 } from 'actions/wallet';
-import {
-    clearData as clearMoonPayData,
-    fetchTransactionDetails as fetchMoonPayTransactionDetails,
-    setAuthenticationStatus as setMoonPayAuthenticationStatus,
-} from 'actions/exchanges/MoonPay';
-import { MOONPAY_RETURN_URL } from 'exchanges/MoonPay';
-import { __DEV__ } from 'config';
 
 import { updateTheme } from 'actions/settings';
 import { fetchNodeList } from 'actions/polling';
 import { dismissAlert, generateAlert } from 'actions/alerts';
-
-import MoonPayKeychainAdapter from 'libs/MoonPay';
 
 import Theme from 'ui/global/Theme';
 import Idle from 'ui/global/Idle';
@@ -58,7 +45,6 @@ import Onboarding from 'ui/views/onboarding/Index';
 import Wallet from 'ui/views/wallet/Index';
 import Settings from 'ui/views/settings/Index';
 import Ledger from 'ui/global/seedStore/Ledger';
-import MoonPayExchange from 'ui/views/exchanges/MoonPay/Index';
 
 /**
  * Wallet wrapper component
@@ -108,8 +94,6 @@ class App extends React.Component {
         /** @ignore */
         deepLinking: PropTypes.bool.isRequired,
         /** @ignore */
-        clearMoonPayData: PropTypes.func.isRequired,
-        /** @ignore */
         forceUpdate: PropTypes.func.isRequired,
         /** @ignore */
         displayTestWarning: PropTypes.func.isRequired,
@@ -122,11 +106,9 @@ class App extends React.Component {
         /** @ignore */
         setDeepLinkContent: PropTypes.func.isRequired,
         /** @ignore */
-        setMoonPayAuthenticationStatus: PropTypes.func.isRequired,
+        displaySeedMigrationAlert: PropTypes.func.isRequired,
         /** @ignore */
-        fetchMoonPayTransactionDetails: PropTypes.func.isRequired,
-        /** @ignore */
-        setViewingMoonpayPurchases: PropTypes.func.isRequired,
+        alerts: PropTypes.bool.isRequired,
     };
 
     constructor(props) {
@@ -152,6 +134,7 @@ class App extends React.Component {
 
         this.checkVaultAvailability();
         this.versionCheck();
+        this.seedMigrationCheck();
     }
 
     componentWillReceiveProps(nextProps) {
@@ -207,21 +190,6 @@ class App extends React.Component {
      */
     setDeepUrl(data) {
         const { deepLinking, generateAlert, t } = this.props;
-
-        const transactionId = last(split(head(split(data, '&')), '='));
-
-        if (data.includes(MOONPAY_RETURN_URL) && isString(transactionId)) {
-            // Check if wallet is ready i.e., user has logged in.
-            // If not, then redirect user to login screen
-            if (this.props.wallet.ready === true) {
-                this.props.fetchMoonPayTransactionDetails(transactionId);
-                return this.props.history.push('/exchanges/moonpay/payment-pending');
-            }
-
-            // Make the MoonPay purchase tab active, so the users can see their new transaction
-            this.props.setViewingMoonpayPurchases(true);
-            return this.props.history.push('/onboarding/');
-        }
 
         this.props.initiateDeepLinkRequest();
         if (!deepLinking) {
@@ -284,6 +252,23 @@ class App extends React.Component {
     }
 
     /**
+     * TEMPORARY: Checks if seed migration tool is up
+     * @param {object} store
+     *
+     * @returns {Promise<object>}
+     *
+     */
+    seedMigrationCheck() {
+        return fetchIsSeedMigrationUp()
+            .then(({ up }) => {
+                if (up.match(VALID_IOTA_SUBDOMAIN_REGEX)) {
+                    this.props.displaySeedMigrationAlert(up);
+                }
+            })
+            .catch(() => {});
+    }
+
+    /**
      * Switch to an account based on account name
      * @param {string} accountName - target account name
      */
@@ -318,29 +303,16 @@ class App extends React.Component {
                 this.props.history.push('/onboarding/seed-intro');
                 break;
             case 'logout':
-                MoonPayKeychainAdapter.clear()
-                    .then(() => {
-                        this.props.clearMoonPayData();
-                        this.props.setMoonPayAuthenticationStatus(false);
-                        this.props.clearWalletData();
-                        this.props.setPassword({});
-                        this.props.setAccountInfoDuringSetup({
-                            name: '',
-                            meta: {},
-                            completed: false,
-                            usedExistingSeed: false,
-                        });
-                        Electron.setOnboardingSeed(null);
-                        this.props.history.push('/onboarding/login');
-                    })
-                    .catch((error) => {
-                        if (__DEV__) {
-                            /* eslint-disable no-console */
-                            console.log(error);
-                            /* eslint-enable no-console */
-                        }
-                    });
-
+                this.props.clearWalletData();
+                this.props.setPassword({});
+                this.props.setAccountInfoDuringSetup({
+                    name: '',
+                    meta: {},
+                    completed: false,
+                    usedExistingSeed: false,
+                });
+                Electron.setOnboardingSeed(null);
+                this.props.history.push('/onboarding/login');
                 break;
             default:
                 if (item.indexOf('settings/account') === 0) {
@@ -388,7 +360,6 @@ class App extends React.Component {
                                 />
                                 <Route path="/wallet" component={Wallet} />
                                 <Route path="/onboarding" component={Onboarding} />
-                                <Route path="/exchanges/moonpay" component={MoonPayExchange} />
                                 <Route loop={false} component={this.Init} />
                             </Switch>
                         </div>
@@ -411,10 +382,10 @@ const mapStateToProps = (state) => ({
     deepLinking: state.settings.deepLinking,
     isBusy:
         !state.wallet.ready || state.ui.isSyncing || state.ui.isSendingTransfer || state.ui.isGeneratingReceiveAddress,
+    alerts: state.alerts,
 });
 
 const mapDispatchToProps = {
-    clearMoonPayData,
     clearWalletData,
     setPassword,
     initiateDeepLinkRequest,
@@ -429,9 +400,7 @@ const mapDispatchToProps = {
     shouldUpdate,
     forceUpdate,
     displayTestWarning,
-    setMoonPayAuthenticationStatus,
-    fetchMoonPayTransactionDetails,
-    setViewingMoonpayPurchases,
+    displaySeedMigrationAlert,
 };
 
 export default withRouter(connect(mapStateToProps, mapDispatchToProps)(withTranslation()(App)));