+++ src/mobile/src/ui/views/wallet/Loading.js
@@ -1,5 +1,3 @@
-import get from 'lodash/get';
-import isNull from 'lodash/isNull';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, View, Text } from 'react-native';
@@ -14,15 +12,8 @@ import { getAccountInfo, getFullAccountInfo } from 'shared-modules/actions/accou
 import { setLoginRoute } from 'shared-modules/actions/ui';
 import { getThemeFromState } from 'shared-modules/selectors/global';
 import { fetchMarketData } from 'shared-modules/actions/polling';
-import {
-    fetchCountries as fetchMoonPayCountries,
-    fetchCurrencies as fetchMoonPayCurrencies,
-    checkIPAddress,
-    refreshCredentialsAndFetchMeta,
-} from 'shared-modules/actions/exchanges/MoonPay';
 import { setSetting } from 'shared-modules/actions/wallet';
 import { changeHomeScreenRoute } from 'shared-modules/actions/home';
-import { __DEV__ } from 'shared-modules/config';
 import {
     getSelectedAccountName,
     getSelectedAccountMeta,
@@ -31,7 +22,6 @@ import {
 } from 'shared-modules/selectors/accounts';
 import { Styling } from 'ui/theme/general';
 import SeedStore from 'libs/SeedStore';
-import { MoonPayKeychainAdapter } from 'libs/keychain';
 import { isAndroid, isIPhoneX } from 'libs/device';
 import { leaveNavigationBreadcrumb } from 'libs/bugsnag';
 import SingleFooterButton from 'ui/components/SingleFooterButton';
@@ -123,14 +113,6 @@ class Loading extends Component {
         deepLinkRequestActive: PropTypes.bool.isRequired,
         /** @ignore */
         setLoginRoute: PropTypes.func.isRequired,
-        /** @ignore */
-        fetchMoonPayCountries: PropTypes.func.isRequired,
-        /** @ignore */
-        fetchMoonPayCurrencies: PropTypes.func.isRequired,
-        /** @ignore */
-        checkIPAddress: PropTypes.func.isRequired,
-        /** @ignore */
-        refreshCredentialsAndFetchMeta: PropTypes.func.isRequired,
         /** All stored account names */
         accountNames: PropTypes.array.isRequired,
         /** @ignore */
@@ -172,7 +154,6 @@ class Loading extends Component {
         }
 
         this.getWalletData();
-        this.fetchMoonPayData();
 
         if (!deepLinkRequestActive) {
             this.props.setSetting('mainSettings');
@@ -229,37 +210,6 @@ class Loading extends Component {
     onChangeNodePress() {
         this.props.setLoginRoute('nodeSettings');
         this.redirectToLogin();
-    }
-
-    /**
-     * Fetches MoonPay data from their servers
-     *
-     * @method fetchMoonPayData
-     *
-     * @returns {void}
-     */
-    fetchMoonPayData() {
-        this.props.fetchMoonPayCountries();
-        this.props.fetchMoonPayCurrencies();
-        this.props.checkIPAddress();
-
-        MoonPayKeychainAdapter.get()
-            .then((credentials) => {
-                if (!isNull(credentials)) {
-                    this.props.refreshCredentialsAndFetchMeta(
-                        get(credentials, 'jwt'),
-                        get(credentials, 'csrfToken'),
-                        MoonPayKeychainAdapter,
-                    );
-                }
-            })
-            .catch((error) => {
-                if (__DEV__) {
-                    /* eslint-disable no-console */
-                    console.log(error);
-                    /* eslint-enable no-console */
-                }
-            });
     }
 
     getWalletData() {
@@ -404,10 +354,6 @@ const mapDispatchToProps = {
     getFullAccountInfo,
     fetchMarketData,
     setLoginRoute,
-    fetchMoonPayCountries,
-    fetchMoonPayCurrencies,
-    checkIPAddress,
-    refreshCredentialsAndFetchMeta,
 };
 
 export default withTranslation(['loading', 'global'])(connect(mapStateToProps, mapDispatchToProps)(Loading));