+++ src/shared/__tests__/reducers/polling.spec.js
@@ -6,7 +6,7 @@ import { PollingActionTypes } from '../../types';
 describe('Reducer: polling', () => {
     it('should have an initial state', () => {
         const initialState = {
-            allPollingServices: ['promotion', 'broadcast', 'marketData', 'nodeList', 'accountInfo', 'moonpayMeta'],
+            allPollingServices: ['promotion', 'broadcast', 'marketData', 'nodeList', 'accountInfo'],
             pollFor: 'promotion',
             retryCount: 0,
             isFetchingPrice: false,
@@ -15,7 +15,6 @@ describe('Reducer: polling', () => {
             isFetchingMarketData: false,
             isFetchingAccountInfo: false,
             isAutoPromoting: false,
-            isFetchingMoonPayMeta: false,
         };
 
         expect(reducer(undefined, {})).to.eql(initialState);