+++ src/shared/reducers/wallet.js
@@ -59,6 +59,14 @@ const initialState = {
      * Determines whether to display test version warning
      */
     displayTestWarning: false,
+    /**
+     * Determines whether to display seed migration alert
+     */
+    displaySeedMigrationAlert: false,
+    /**
+     * Seed migration url
+     */
+    seedMigrationUrl: '',
 };
 
 export default (state = initialState, action) => {
@@ -206,6 +214,12 @@ export default (state = initialState, action) => {
             return {
                 ...state,
                 displayTestWarning: true,
+            };
+        case WalletActionTypes.DISPLAY_SEED_MIGRATION_ALERT:
+            return {
+                ...state,
+                displaySeedMigrationAlert: true,
+                seedMigrationUrl: action.payload,
             };
         default:
             return state;